Push paging logic into Ldp::Response
