+++ lib/ldp/response.rb
@@ -3,9 +3,6 @@ module Ldp
     extend Forwardable
     def_delegators :@response, :body, :headers, :env, :success?, :status
 
-    require 'ldp/response/paging'
-    include Ldp::Response::Paging
-
     TYPE = 'type'.freeze
 
     attr_reader :response
@@ -79,7 +76,11 @@ module Ldp
     ##
     # Get the subject for the response
     def subject
-      page_subject
+      @subject ||= if has_page?
+        graph.first_object [page_subject, RDF::Vocab::LDP.pageOf, nil]
+      else
+        page_subject
+      end
     end
 
     ##
@@ -142,6 +143,47 @@ module Ldp
     def minimal?
       preferences[RETURN][:value] == "minimal"
     end
+
+    ##
+    # Statements about the page
+    def page
+      @page_graph ||= begin
+        g = RDF::Graph.new  
+
+        if resource?
+          res = graph.query RDF::Statement.new(page_subject, nil, nil)
+
+          res.each_statement do |s|
+            g << s
+          end
+        end
+
+        g
+      end
+    end
+
+    ##
+    # Is there a next page?
+    def has_next?
+      next_page != nil
+    end
+
+    ##
+    # Get the URI for the next page
+    def next_page
+      graph.first_object [page_subject, RDF::Vocab::LDP.nextPage, nil]
+    end
+
+    ##
+    # Get the URI to the first page
+    def first_page
+      if links['first']
+        RDF::URI.new links['first']
+      elsif graph.has_statement? RDf::Statement.new(page_subject, RDF::Vocab::LDP.nextPage, nil)
+        subject
+      end
+    end
+
 
     private
